requirements_gpu.txt
tensorflow-gpu==2.10.0
numpy>=1.22.0,<1.24.0
pandas>=1.4.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
scipy>=1.7.0
seaborn>=0.11.0
Flask>=2.0.0
Werkzeug>=2.0.0
Pillow>=9.0.0
protobuf==3.19.6
